Mmmmmm good first draft of the documentation. Still need to add like a quickstart tutorial and the build instructions. Not that they're that involved through.... The docs are a little wonky in this draft, but I'll weed out the... weeds. Heh. It is 2:26 in the AM now though - been a while since I've been up this late.
It's funny: I remember that a year ago, around this time, I was desparing over a quiz making framework.
I had decided that I would make a custom programming language for it, and create a BYTECODE COMPILER for it...
Yah, I've come kinda' far if you think about it.
I also remember that I watched Kimi No Na Wa on that night - it's really been a year since I last watched it...
Damn...